+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * ADD：minitable onFilter支持默认处理
 * FIX: minitable restFun
 * FIX：minitable 选择列点击问题
+* FIX: SipRestSqlType.List pageSize = 999
 
 ## 0.2.4